requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cfgv==3.4.0
click==8.1.6
distlib==0.3.7
fastapi==0.101.0
filelock==3.12.2
h11==0.14.0
httptools==0.6.0
identify==2.5.27
idna==3.4
iniconfig==2.0.0
nodeenv==1.8.0
packaging==23.1
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
psycopg2-binary==2.9.7
py4j==0.10.9.5
pydantic==2.1.1
pydantic_core==2.4.0
pyspark==3.3.0
pytest==7.4.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.24.3
watchfiles==0.19.0
websockets==11.0.3
